requirements.txt
python-dotenv
slack_bolt
google-generativeai
chromadb